Fixed scythe_get_key to not error.
When keys were missing, scythe_get_key would throw an error, causing tests to fail as described in issue #16. The refactored method takes a "source" parameter and returns the secret value of that single key, or NA. The version that returned mutliple hardcoded key values was not needed (keyring::list_keys() works better), and the test that used it was refactored.